requirements.txt
fastapi
uvicorn
starlette
strawberry-graphql[fastapi]==0.217.0
sqlalchemy
psycopg2-binary
shapely
requests
pydantic[email]
python-dotenv
boto3
jinja2
httpx
aiofiles
fiona
rasterio
pyproj
pandas
openpyxl